refactor: 훅 폴더구조 변경
